Develop

+++ task_management/src/main/java/com/example/entity/TimersSetting.java
@@ -22,10 +22,10 @@ public class TimersSetting {
 	@SequenceGenerator(name = "TIMERS_SETTINGS_ID_GENERATOR", sequenceName = "TIMERS_SETTINGS_ID_SEQ", allocationSize = 1)
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "TIMERS_SETTINGS_ID_GENERATOR")
 	@Column(name = "ID")
-	private Integer id; // タイマー設定ID
+	private Long id; // タイマー設定ID
 	
 	@Column(name = "USER_ID")
-	private Integer userId; // ユーザーID(外部キー)
+	private Long userId; // ユーザーID(外部キー)
 	
 	@Column(name = "FOCUS_TIME")
 	private String focusTime; // 集中時間
@@ -40,21 +40,21 @@ public class TimersSetting {
 	// アクセサメソッド
 	
 	// タイマー設定ID
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 	
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	
 	
 	// ユーザーID
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 	
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
